Cleanup: Remove Default*ChannelConfig implementations which are not used anymore
Motivation:

We have various Default*ChannelConfig classes that are unused and so should be removed

Modifications:

Remove unused classes

Result:

Cleanup
