+++ handler/src/test/java/io/netty/handler/pcap/PcapWriteHandlerTest.java
@@ -18,6 +18,7 @@ package io.netty.handler.pcap;
 import io.netty.bootstrap.Bootstrap;
 import io.netty.bootstrap.ServerBootstrap;
 import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufAllocator;
 import io.netty.buffer.ByteBufOutputStream;
 import io.netty.buffer.ByteBufUtil;
 import io.netty.buffer.Unpooled;
@@ -30,16 +31,19 @@ import io.netty.channel.ChannelOption;
 import io.netty.channel.ChannelOutboundHandlerAdapter;
 import io.netty.channel.ChannelPipeline;
 import io.netty.channel.ChannelPromise;
+import io.netty.channel.DefaultChannelConfig;
 import io.netty.channel.DefaultChannelId;
 import io.netty.channel.EventLoopGroup;
+import io.netty.channel.MessageSizeEstimator;
 import io.netty.channel.MultiThreadIoEventLoopGroup;
+import io.netty.channel.RecvByteBufAllocator;
 import io.netty.channel.SimpleChannelInboundHandler;
+import io.netty.channel.WriteBufferWaterMark;
 import io.netty.channel.embedded.EmbeddedChannel;
 import io.netty.channel.nio.NioIoHandler;
 import io.netty.channel.socket.DatagramChannel;
 import io.netty.channel.socket.DatagramChannelConfig;
 import io.netty.channel.socket.DatagramPacket;
-import io.netty.channel.socket.DefaultDatagramChannelConfig;
 import io.netty.channel.socket.SocketChannel;
 import io.netty.channel.socket.nio.NioDatagramChannel;
 import io.netty.channel.socket.nio.NioServerSocketChannel;
@@ -51,7 +55,6 @@ import org.junit.jupiter.api.Test;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
-import java.net.DatagramSocket;
 import java.net.Inet4Address;
 import java.net.InetAddress;
 import java.net.InetSocketAddress;
@@ -1358,11 +1361,7 @@ public class PcapWriteHandlerTest {
         public DatagramChannelConfig config() {
             if (config == null) {
                 // ick! config() is called by the super constructor, so we need to do this.
-                try {
-                    config = new DefaultDatagramChannelConfig(this, new DatagramSocket());
-                } catch (SocketException e) {
-                    throw new RuntimeException(e);
-                }
+                config = new DatagramChannelConfigImpl(this);
             }
             return config;
         }
@@ -1504,6 +1503,176 @@ public class PcapWriteHandlerTest {
             //Discard
             ReferenceCountUtil.release(msg);
             promise.setSuccess();
+        }
+    }
+
+    private static final class DatagramChannelConfigImpl extends DefaultChannelConfig
+            implements DatagramChannelConfig {
+        DatagramChannelConfigImpl(Channel channel) {
+            super(channel);
+        }
+
+        @Override
+        public int getSendBufferSize() {
+            return 0;
+        }
+
+        @Override
+        public DatagramChannelConfig setSendBufferSize(int sendBufferSize) {
+            return this;
+        }
+
+        @Override
+        public int getReceiveBufferSize() {
+            return 0;
+        }
+
+        @Override
+        public DatagramChannelConfig setReceiveBufferSize(int receiveBufferSize) {
+            return this;
+        }
+
+        @Override
+        public int getTrafficClass() {
+            return 0;
+        }
+
+        @Override
+        public DatagramChannelConfig setTrafficClass(int trafficClass) {
+            return this;
+        }
+
+        @Override
+        public boolean isReuseAddress() {
+            return false;
+        }
+
+        @Override
+        public DatagramChannelConfig setReuseAddress(boolean reuseAddress) {
+            return this;
+        }
+
+        @Override
+        public boolean isBroadcast() {
+            return false;
+        }
+
+        @Override
+        public DatagramChannelConfig setBroadcast(boolean broadcast) {
+            return this;
+        }
+
+        @Override
+        public boolean isLoopbackModeDisabled() {
+            return false;
+        }
+
+        @Override
+        public DatagramChannelConfig setLoopbackModeDisabled(boolean loopbackModeDisabled) {
+            return this;
+        }
+
+        @Override
+        public int getTimeToLive() {
+            return 0;
+        }
+
+        @Override
+        public DatagramChannelConfig setTimeToLive(int ttl) {
+            return this;
+        }
+
+        @Override
+        public InetAddress getInterface() {
+            return null;
+        }
+
+        @Override
+        public DatagramChannelConfig setInterface(InetAddress interfaceAddress) {
+            return this;
+        }
+
+        @Override
+        public NetworkInterface getNetworkInterface() {
+            return null;
+        }
+
+        @Override
+        public DatagramChannelConfig setNetworkInterface(NetworkInterface networkInterface) {
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setWriteSpinCount(int writeSpinCount) {
+            super.setWriteSpinCount(writeSpinCount);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setConnectTimeoutMillis(int connectTimeoutMillis) {
+            super.setConnectTimeoutMillis(connectTimeoutMillis);
+            return this;
+        }
+
+        @Override
+        @Deprecated
+        public DatagramChannelConfig setMaxMessagesPerRead(int maxMessagesPerRead) {
+            super.setMaxMessagesPerRead(maxMessagesPerRead);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setAllocator(ByteBufAllocator allocator) {
+            super.setAllocator(allocator);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setRecvByteBufAllocator(RecvByteBufAllocator allocator) {
+            super.setRecvByteBufAllocator(allocator);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setAutoRead(boolean autoRead) {
+            super.setAutoRead(autoRead);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setAutoClose(boolean autoClose) {
+            super.setAutoClose(autoClose);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setWriteBufferHighWaterMark(int writeBufferHighWaterMark) {
+            super.setWriteBufferHighWaterMark(writeBufferHighWaterMark);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setWriteBufferLowWaterMark(int writeBufferLowWaterMark) {
+            super.setWriteBufferLowWaterMark(writeBufferLowWaterMark);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setWriteBufferWaterMark(WriteBufferWaterMark writeBufferWaterMark) {
+            super.setWriteBufferWaterMark(writeBufferWaterMark);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setMessageSizeEstimator(MessageSizeEstimator estimator) {
+            super.setMessageSizeEstimator(estimator);
+            return this;
+        }
+
+        @Override
+        public DatagramChannelConfig setMaxMessagesPerWrite(int maxMessagesPerWrite) {
+            super.setMaxMessagesPerWrite(maxMessagesPerWrite);
+            return this;
         }
     }
 }
