--- transport/src/main/java/io/netty/channel/socket/DefaultServerSocketChannelConfig.java
@@ -1,209 +1,0 @@
-/*
- * Copyright 2012 The Netty Project
- *
- * The Netty Project licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-package io.netty.channel.socket;
-
-import io.netty.buffer.ByteBufAllocator;
-import io.netty.channel.ChannelException;
-import io.netty.channel.ChannelOption;
-import io.netty.channel.DefaultChannelConfig;
-import io.netty.channel.MessageSizeEstimator;
-import io.netty.channel.RecvByteBufAllocator;
-import io.netty.channel.ServerChannelRecvByteBufAllocator;
-import io.netty.channel.WriteBufferWaterMark;
-import io.netty.util.NetUtil;
-import io.netty.util.internal.ObjectUtil;
-
-import java.net.ServerSocket;
-import java.net.SocketException;
-import java.util.Map;
-
-import static io.netty.channel.ChannelOption.SO_BACKLOG;
-import static io.netty.channel.ChannelOption.SO_RCVBUF;
-import static io.netty.channel.ChannelOption.SO_REUSEADDR;
-import static io.netty.util.internal.ObjectUtil.checkPositiveOrZero;
-
-/**
- * The default {@link ServerSocketChannelConfig} implementation.
- */
-public class DefaultServerSocketChannelConfig extends DefaultChannelConfig
-                                              implements ServerSocketChannelConfig {
-
-    protected final ServerSocket javaSocket;
-    private volatile int backlog = NetUtil.SOMAXCONN;
-
-    /**
-     * Creates a new instance.
-     */
-    public DefaultServerSocketChannelConfig(ServerSocketChannel channel, ServerSocket javaSocket) {
-        super(channel, new ServerChannelRecvByteBufAllocator());
-        this.javaSocket = ObjectUtil.checkNotNull(javaSocket, "javaSocket");
-    }
-
-    @Override
-    public Map<ChannelOption<?>, Object> getOptions() {
-        return getOptions(super.getOptions(), SO_RCVBUF, SO_REUSEADDR, SO_BACKLOG);
-    }
-
-    @SuppressWarnings("unchecked")
-    @Override
-    public <T> T getOption(ChannelOption<T> option) {
-        if (option == SO_RCVBUF) {
-            return (T) Integer.valueOf(getReceiveBufferSize());
-        }
-        if (option == SO_REUSEADDR) {
-            return (T) Boolean.valueOf(isReuseAddress());
-        }
-        if (option == SO_BACKLOG) {
-            return (T) Integer.valueOf(getBacklog());
-        }
-
-        return super.getOption(option);
-    }
-
-    @Override
-    public <T> boolean setOption(ChannelOption<T> option, T value) {
-        validate(option, value);
-
-        if (option == SO_RCVBUF) {
-            setReceiveBufferSize((Integer) value);
-        } else if (option == SO_REUSEADDR) {
-            setReuseAddress((Boolean) value);
-        } else if (option == SO_BACKLOG) {
-            setBacklog((Integer) value);
-        } else {
-            return super.setOption(option, value);
-        }
-
-        return true;
-    }
-
-    @Override
-    public boolean isReuseAddress() {
-        try {
-            return javaSocket.getReuseAddress();
-        } catch (SocketException e) {
-            throw new ChannelException(e);
-        }
-    }
-
-    @Override
-    public ServerSocketChannelConfig setReuseAddress(boolean reuseAddress) {
-        try {
-            javaSocket.setReuseAddress(reuseAddress);
-        } catch (SocketException e) {
-            throw new ChannelException(e);
-        }
-        return this;
-    }
-
-    @Override
-    public int getReceiveBufferSize() {
-        try {
-            return javaSocket.getReceiveBufferSize();
-        } catch (SocketException e) {
-            throw new ChannelException(e);
-        }
-    }
-
-    @Override
-    public ServerSocketChannelConfig setReceiveBufferSize(int receiveBufferSize) {
-        try {
-            javaSocket.setReceiveBufferSize(receiveBufferSize);
-        } catch (SocketException e) {
-            throw new ChannelException(e);
-        }
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setPerformancePreferences(int connectionTime, int latency, int bandwidth) {
-        javaSocket.setPerformancePreferences(connectionTime, latency, bandwidth);
-        return this;
-    }
-
-    @Override
-    public int getBacklog() {
-        return backlog;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setBacklog(int backlog) {
-        checkPositiveOrZero(backlog, "backlog");
-        this.backlog = backlog;
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setConnectTimeoutMillis(int connectTimeoutMillis) {
-        super.setConnectTimeoutMillis(connectTimeoutMillis);
-        return this;
-    }
-
-    @Override
-    @Deprecated
-    public ServerSocketChannelConfig setMaxMessagesPerRead(int maxMessagesPerRead) {
-        super.setMaxMessagesPerRead(maxMessagesPerRead);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setWriteSpinCount(int writeSpinCount) {
-        super.setWriteSpinCount(writeSpinCount);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setAllocator(ByteBufAllocator allocator) {
-        super.setAllocator(allocator);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setRecvByteBufAllocator(RecvByteBufAllocator allocator) {
-        super.setRecvByteBufAllocator(allocator);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setAutoRead(boolean autoRead) {
-        super.setAutoRead(autoRead);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setWriteBufferHighWaterMark(int writeBufferHighWaterMark) {
-        super.setWriteBufferHighWaterMark(writeBufferHighWaterMark);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setWriteBufferLowWaterMark(int writeBufferLowWaterMark) {
-        super.setWriteBufferLowWaterMark(writeBufferLowWaterMark);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setWriteBufferWaterMark(WriteBufferWaterMark writeBufferWaterMark) {
-        super.setWriteBufferWaterMark(writeBufferWaterMark);
-        return this;
-    }
-
-    @Override
-    public ServerSocketChannelConfig setMessageSizeEstimator(MessageSizeEstimator estimator) {
-        super.setMessageSizeEstimator(estimator);
-        return this;
-    }
-}
